Add basic support for having multiple instances per server

+++ umbrellapanel-chief/src/main/java/ms/idrea/umbrellapanel/chief/UmbrellaChief.java
@@ -6,18 +6,23 @@ import java.util.logging.Level;
 import java.util.logging.Logger;
 
 import lombok.Getter;
+
 import ms.idrea.umbrellapanel.api.chief.Chief;
 import ms.idrea.umbrellapanel.api.chief.PanelUserDatabase;
 import ms.idrea.umbrellapanel.api.chief.WorkerManager;
 import ms.idrea.umbrellapanel.api.chief.conf.ChiefProperties;
-import ms.idrea.umbrellapanel.api.chief.gameserver.GameServer;
-import ms.idrea.umbrellapanel.api.chief.gameserver.ServerManager;
 import ms.idrea.umbrellapanel.api.chief.net.NetworkServer;
 import ms.idrea.umbrellapanel.api.chief.net.RunningWorker;
 import ms.idrea.umbrellapanel.api.chief.webapi.EndPointManager;
 import ms.idrea.umbrellapanel.api.core.permissions.PanelUser;
+import ms.idrea.umbrellapanel.api.gameserver.ManagedServer;
+import ms.idrea.umbrellapanel.api.gameserver.MultiInstanceServer;
+import ms.idrea.umbrellapanel.api.gameserver.ServerInstance;
+import ms.idrea.umbrellapanel.api.gameserver.ServerManager;
+import ms.idrea.umbrellapanel.api.gameserver.SingleInstanceServer;
 import ms.idrea.umbrellapanel.api.util.Address;
 import ms.idrea.umbrellapanel.api.util.LoggerHelper;
+import ms.idrea.umbrellapanel.api.worker.gameserver.GameServer;
 import ms.idrea.umbrellapanel.chief.conf.UmbrellaChiefProperties;
 import ms.idrea.umbrellapanel.chief.gameserver.UmbrellaServerManager;
 import ms.idrea.umbrellapanel.chief.net.UmbrellaNetworkServer;
@@ -40,7 +45,8 @@ public class UmbrellaChief implements Chief {
 		instance.enableConsole();
 	}
 
-	private UmbrellaChief() {}
+	private UmbrellaChief() {
+	}
 
 	private ChiefProperties chiefProperties;
 	private FileManager fileManager;
@@ -82,7 +88,7 @@ public class UmbrellaChief implements Chief {
 		fileManager.register(panelUserDatabase, "users");
 		fileManager.load();
 	}
-	
+
 	public void enableConsole() {
 		Scanner scanner = new Scanner(System.in);
 		String line;
@@ -119,21 +125,50 @@ public class UmbrellaChief implements Chief {
 				PanelUser user = panelUserDatabase.createUser(args[0], args[1]);
 				System.out.println(user);
 			} else if (base.equalsIgnoreCase("createserver")) {
-				GameServer server = serverManager.createServer(new Address(args[0], Integer.valueOf(args[1])), "java -jar server.jar", Integer.valueOf(args[2]));
+				GameServer server = serverManager.createSingleInstanceServer(new Address(args[0], Integer.valueOf(args[1])), "java -jar server.jar", Integer.valueOf(args[2]));
 				System.out.println(server);
+			} else if (base.equalsIgnoreCase("createmultiserver")) {
+				GameServer server = serverManager.createMultiInstanceServer("java -jar server.jar", Integer.valueOf(args[0]));
+				System.out.println(server);
+			} else if (base.equalsIgnoreCase("createinstance")) {
+				MultiInstanceServer server = (MultiInstanceServer) serverManager.getServer(Integer.valueOf(args[0]));
+				ServerInstance instance = server.createNewInstance(-1, new Address(args[1], Integer.valueOf(args[2])));
+				System.out.println(instance);
 			} else if (base.equalsIgnoreCase("manageserver")) {
-				GameServer server = serverManager.getServer(Integer.valueOf(args[1]));
+				ManagedServer server = serverManager.getServer(Integer.valueOf(args[1]));
+				if (args[0].equalsIgnoreCase("delete")) {
+					serverManager.deleteServer(server);
+				} else if (server instanceof SingleInstanceServer) {
+					switch (args[0]) {
+						case "start":
+							((SingleInstanceServer) server).start();
+							break;
+						case "force-stop":
+							((SingleInstanceServer) server).forceStop();
+							break;
+						default:
+							System.out.println("not found");
+							break;
+					}
+				} else {
+					System.out.println("Only supported for single instance servers.");
+				}
+				System.out.println("OK");
+			} else if (base.equalsIgnoreCase("manageinstance")) {
+				MultiInstanceServer server = (MultiInstanceServer) serverManager.getServer(Integer.valueOf(args[1]));
+				ServerInstance instance = server.getInstance(Integer.valueOf(args[2]));
 				switch (args[0]) {
 					case "start":
-						server.start();
+						instance.start();
 						break;
 					case "force-stop":
-						server.forceStop();
+						instance.forceStop();
 						break;
 					case "delete":
-						serverManager.deleteServer(server);
+						instance.delete();
 						break;
 					default:
+						System.out.println("not found");
 						break;
 				}
 				System.out.println("OK");
@@ -143,7 +178,7 @@ public class UmbrellaChief implements Chief {
 				for (int i = 1; i < args.length; i++) {
 					s.append(args[i]).append(' ');
 				}
-				server.dispatchCommand(s.substring(0, s.length() - 1));
+				((SingleInstanceServer) server).dispatchCommand(s.substring(0, s.length() - 1));
 				System.out.println("OK");
 			} else if (base.equalsIgnoreCase("listworkers")) {
 				System.out.println("---Workers---");
@@ -176,7 +211,10 @@ public class UmbrellaChief implements Chief {
 				logger.info("commands:");
 				logger.info("adduser <name> <password>");
 				logger.info("crateserver <hostIp> <hostPort> <workerId>");
+				logger.info("createmultiserver <workerId>");
+				logger.info("createinstance <serverId> <hostIp> <hostPort>");
 				logger.info("manageserver <start|force-stop|delete> <serverId>");
+				logger.info("manageinstance <start|force-stop|delete> <serverId> <instanceId>");
 				logger.info("sendcmd <serverId> <cmd...>");
 				logger.info("listworkers");
 				logger.info("listusers");
@@ -184,6 +222,7 @@ public class UmbrellaChief implements Chief {
 				logger.info("addperm <userId> <value>");
 				logger.info("addperm <userId> <serverId> <value>");
 			} else {
+				System.out.println(serverManager);
 				logger.info("Unknown command! For help type \"help\"");
 			}
 		} catch (Exception e) {
